requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.0
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.8.0
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
lxml==4.6.3
psycopg2==2.9.1
python-decouple==3.4
pytz==2021.1
requests==2.26.0
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.6
whitenoise==5.3.0
